move medication category constants to separate folder
