+++ backend/src/constants/MedicationCategory/medsUnits.js
@@ -1,0 +1,2 @@
+export const VOLUME_UNIT = 'ml';
+export const QUANTITY_UNIT = 'pcs';
